+++ app/screens/AppNavigator/screens/Settings/components/RowNetworkItem.tsx
@@ -1,0 +1,66 @@
+import { MaterialIcons } from '@expo/vector-icons'
+import { NavigationProp, useNavigation } from '@react-navigation/native'
+import * as React from 'react'
+import { useCallback } from 'react'
+import { Alert, Platform, TouchableOpacity } from 'react-native'
+import { Text } from '../../../../../components'
+import { useNetworkContext } from '../../../../../contexts/NetworkContext'
+import { EnvironmentNetwork, isPlayground } from '../../../../../environment'
+import { tailwind } from '../../../../../tailwind'
+import { translate } from '../../../../../translations'
+import { SettingsParamList } from '../SettingsNavigator'
+
+export function RowNetworkItem (props: { network: EnvironmentNetwork }): JSX.Element {
+  const navigation = useNavigation<NavigationProp<SettingsParamList>>()
+  const { network, updateNetwork } = useNetworkContext()
+
+  const onPress = useCallback(async () => {
+    if (props.network === network) {
+      if (isPlayground(props.network)) {
+        navigation.navigate('Playground')
+      }
+    } else {
+      if (Platform.OS !== 'web') {
+        Alert.alert(
+          translate('screens/Settings', 'Network Switch'),
+          translate('screens/Settings', `You are about to switch to ${props.network}. If there is no existing wallet on this network, you will be redirected to Onboarding screen. Do you want to proceed?`),
+          [
+            {
+              text: translate('screens/Settings', 'No'),
+              style: 'cancel'
+            },
+            {
+              text: translate('screens/Settings', 'Yes'),
+              style: 'destructive',
+              onPress: async () => await updateNetwork(props.network)
+            }
+          ]
+        )
+      } else {
+        await updateNetwork(props.network)
+      }
+    }
+  }, [network])
+
+  return (
+    <TouchableOpacity
+      testID={`button_network_${props.network}`}
+      style={tailwind('flex flex-row p-4 pr-2 bg-white items-center justify-between border-b border-gray-200')}
+      onPress={onPress}
+    >
+      <Text style={tailwind('font-medium')}>
+        {props.network}
+      </Text>
+
+      {
+        props.network === network &&
+        (
+          <MaterialIcons
+            testID={`button_network_${props.network}_check`} size={24} name='check'
+            style={tailwind('text-primary')}
+          />
+        )
+      }
+    </TouchableOpacity>
+  )
+}
